apps/api/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.2.0
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-exporter-prometheus==0.59b0
httpx==0.28.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.35.0

